Hiding nav bar for now to deploy.
