+++ src/routes/+page.svelte
@@ -1,4 +1,10 @@
 <svelte:head>
-	<title>wesflorence</title>
+  <title>wesflorence</title>
 </svelte:head>
 <h1 class="text-3xl font-bold">Wes Florence</h1>
+<p>Hello. I'm a Software Developer living in Seattle, WA.</p>
+<div class="text-zinc-400">
+<a href="https://github.com/wesleyflorence" class="hover:text-zinc-50">Github</a>
+|
+<a href="https://www.linkedin.com/in/wesflorence" class="hover:text-zinc-50">LinkedIn</a>
+</div>
