+++ src/routes/blog/1.md
@@ -3,9 +3,4 @@ title: Post One
 date: '2023.11.01'
 ---
 
-Hello from _Post One._
-<br>
-
-```js
-let x = hello;
-```
+init.
