--- src/routes/blog/2.md
@@ -1,11 +1,0 @@
----
-title: Post Two
-date: '2023.11.10'
----
-
-Hello from _Post One._
-<br>
-
-```js
-let x = hello;
-```
