--- src/routes/blog/3.md
@@ -1,15 +1,0 @@
----
-title: Post 3
-date: '2023.12.01'
----
-
-## Heading
-
-Hello from _Post One._
-<br>
-
-```js
-let x = hello;
-```
-
-## Heading 2
