+++ src/routes/cancer/+page.svelte
@@ -2,3 +2,4 @@
 	<title>wesflorence :: cancer db</title>
 </svelte:head>
 <h1 class="text-4xl font-bold mb-2">Cancer</h1>
+<p>Dasboard for tracking my Hodgkin Lymphoma</p>
