+++ src/routes/research/+page.svelte
@@ -3,6 +3,4 @@
 </svelte:head>
 
 <h1 class="text-4xl font-bold mb-2">Research</h1>
-Reading list
-<br />
-Wiki
+Reading list and knowledge aggregation.
